modules/module2-questionnaire/backend/pyproject.toml
[project]
name = "synergy-questionnaire-ai"
version = "0.1.0"
description = "Synergy AI Module 2 POC — questionnaire-driven marketing advice for new coaches"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "litellm>=1.52.0",
    "python-dotenv>=1.0.1",
    "openpyxl>=3.1.5",
    "python-docx>=1.1.2",
    "httpx>=0.27.2",
    "tzdata>=2026.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
